refactor - modify menus package structure
- 메뉴와 관련된 패키지 구조를 애그리거트 중심 구조로 변경했습니다
