+++ src/main/java/kitchenpos/menus/model/Menu.java
@@ -1,4 +1,4 @@
-package kitchenpos.model;
+package kitchenpos.menus.model;
 
 import java.math.BigDecimal;
 import java.util.List;
