+++ src/test/java/kitchenpos/bo/MenuBoTest.java
@@ -1,10 +1,11 @@
 package kitchenpos.bo;
 
-import kitchenpos.dao.MenuDao;
+import kitchenpos.menus.dao.MenuDao;
 import kitchenpos.menugroups.dao.MenuGroupDao;
-import kitchenpos.dao.MenuProductDao;
+import kitchenpos.menus.dao.MenuProductDao;
+import kitchenpos.menus.bo.MenuBo;
 import kitchenpos.products.dao.ProductDao;
-import kitchenpos.model.Menu;
+import kitchenpos.menus.model.Menu;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
